+++ .prettierrc.json
@@ -3,5 +3,5 @@
     "useTabs": true,
     "semi": false,
     "singleQuote": true,
-    "trailingComma": "es5"
+    "trailingComma": "all"
 }
