+++ package.json
@@ -28,8 +28,9 @@
 	"devDependencies": {
 		"@types/node": "^20.14.10",
 		"just-group-by": "^2.2.0",
-		"prettier": "^2.8.1",
+		"prettier": "^3.3.3",
 		"remark": "^15.0.1",
+		"remark-directive": "^3.0.0",
 		"remark-mdx": "^3.0.1",
 		"remark-parse": "^11.0.0",
 		"unified": "^11.0.5",
