--- versions/2.0.3/bin/const.js
@@ -1,5 +1,0 @@
-import { dirname } from 'path'
-import { fileURLToPath } from 'url'
-
-export const __filename = fileURLToPath(import.meta.url)
-export const __dirname = dirname(__filename)
